requirements.txt
requests
python-dotenv
colorama
rich
pygments